refactor(*,-tsconfig): baseUrl 제거하고 모듈 임포트 상대 경로로 통일하기
