+++ server/src/problem/problem.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import Problem from 'src/entities/problem.entity';
+import Problem from '../entities/problem.entity';
 import { ProblemController } from './problem.controller';
 import { ProblemService } from './problem.service';
