+++ server/src/room-user/room-user.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import RoomUser from 'src/entities/room-user.entity';
+import RoomUser from '../room-user/room-user.entity';
 import { RoomUserService } from './room-user.service';
 
 @Module({
