+++ server/src/room/room.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import Room from 'src/entities/room.entity';
-import { RoomUserModule } from 'src/room-user/room-user.module';
-import { UserModule } from 'src/user/user.module';
+import Room from '../entities/room.entity';
+import { RoomUserModule } from '../room-user/room-user.module';
+import { UserModule } from '../user/user.module';
 import { RoomController } from './room.controller';
 import { RoomService } from './room.service';
 import { SocketModule } from '../socket/socket.module';
