+++ server/src/socket/socket.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { SocketGateway } from './socket.gateway';
 import { UserModule } from '../user/user.module';
 import { SocketService } from './socket.service';
-import { ProblemModule } from 'src/problem/problem.module';
+import { ProblemModule } from '../problem/problem.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import Room from '../entities/room.entity';
