+++ server/src/submission/submission.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import Submission from 'src/entities/submission.entity';
-import { ProblemModule } from 'src/problem/problem.module';
-import { RoomUserModule } from 'src/room-user/room-user.module';
-import { UserModule } from 'src/user/user.module';
+import Submission from '../entities/submission.entity';
+import { ProblemModule } from '../problem/problem.module';
+import { RoomUserModule } from '../room-user/room-user.module';
+import { UserModule } from '../user/user.module';
 import { SubmissionController } from './submission.controller';
 import { SubmissionService } from './submission.service';
 import { SocketModule } from '../socket/socket.module';
-import { RoomModule } from 'src/room/room.module';
+import { RoomModule } from '../room/room.module';
 
 @Module({
   imports: [
